refactor app.py more - move honeycomb, rollbar, xray and cloudwatch to their own files
